requirements.txt
langchain
langchain-openai
langchain-core
python-dotenv
pydantic